Fix YouTube embeds not starting at the correct time (#27)
Sometimes YouTube videos don't start at the beginning. This seems
to occurs under the following conditions:

- The user is signed into YouTube (Google)
- The user has enabled (not disabled) YouTube video watch history
- The user watches a YouTube video for some minimum amount of time

When the above factors are present, anytime the user plays that
particular video, whether on YouTube or as an embed, it will
start playing at the last saved time. This appears to be done as a
convenience for users.

There doesn't seem to be a way of disabling this feature. Since this
is not behavior we would like for this site, we instead perform the
following compensation:

- If a recording is not "live", embed and start playing the video.
- Star the video muted so as to not cause a brief sound blip.
- After it starts playing, seek to the beginning (or skip time).
- Unmute the video.